+++ NEWS.md
@@ -1,3 +1,19 @@
+# plotannotate 1.3-0 (2025-05-29)
+
+- improved the circle drawing code, so that either the x- or y-axis distance properly determines the circle radius
+
+- added an ellipse drawing mode (makes use of the `conicfit` package)
+
+- added a smooth mode for freehand drawing
+
+- much faster undo for freehand drawing
+
+# plotannotate 1.2-0 (2025-05-22)
+
+- minor documentation updates
+
+- use `on.exit()` as needed to reset `par()` changes
+
 # plotannotate 1.0-0 (2025-05-19)
 
 - first release